Fix race condition when invoiced in parallel
When the exiftool based loaded is invoiced twice for the same image, a
race condition lead to a crash.
The parallel execution happens when the main image as well as the
thumbnail manager tries to load the same image.
This is mitigated by adding a unique timestamp to the temporary path